refactor: gameContainer
